+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- `DatePickerField` is less prone to trigger the `onChangeValue` event if not needed
+
 ## [v1.1.0] - 2021-05-24
 
 ### Added